chore: rename sub-package name for view and UI from 'view' to 'ui' to make the domain and ui separation clear
